[#31]/feat atlaskitTreeFinder 함수 구현
